[fix] 스토리 서비스와 스토리 댓글 서비스 분류
